I have terminated dto's structure, and i've done services structures. Obviusly evrerything can be changed.

+++ src/main/java/srl/narrel/demo/dto/GarageDTO.java
@@ -15,6 +15,6 @@ public class GarageDTO {
 	//ClassDto inherits father's field
 	private CondominioModel condominio;
 	private int numero;
-	private double consumoElettrico;
+	private float consumoElettrico;
 	
 }
